[refactor] getInstanceResponseDto() 함수를 생성자로 변경
