+++ src/instances/dto/instance-response.dto.ts
@@ -1,4 +1,5 @@
 import { ApiProperty } from '@nestjs/swagger';
+import { InstanceModel } from 'src/instances/instances.service';
 
 export enum InstanceType {
   T2MICRO = 't2.micro',
@@ -90,4 +91,30 @@ export class InstanceResponseDto {
     type: InstanceInformations,
   })
   informations: InstanceInformations;
+
+  constructor(savedInstance: InstanceModel, fetchedInstance: AWS.EC2.Instance) {
+    const instanceOption: InstanceOption = {
+      name: savedInstance.name,
+    };
+
+    const instanceInformations: InstanceInformations = {
+      id: savedInstance.instanceId,
+      type: savedInstance.type,
+      os: savedInstance.os,
+      tier: savedInstance.tier,
+      publicIp: fetchedInstance.PublicIpAddress || null,
+      privateIp: fetchedInstance.PrivateIpAddress || null,
+      securityGroup: [
+        'tcp : 80 - 0.0.0.0/0',
+        'tcp : 22 - 0.0.0.0/0',
+        'tcp : 443 - 0.0.0.0/0',
+      ],
+    };
+    const response: InstanceResponseDto = {
+      options: instanceOption,
+      informations: instanceInformations,
+    };
+
+    return response;
+  }
 }
