+++ src/instances/utils/getInstances.ts
@@ -1,9 +1,5 @@
 import * as AWS from 'aws-sdk';
 import { InstanceResponseDto } from 'src/instances/dto';
-import {
-  InstanceInformations,
-  InstanceOption,
-} from 'src/instances/dto/instance-response.dto';
 import { InstanceModel } from 'src/instances/instances.service';
 
 export const getNameFromTag = (
@@ -21,35 +17,6 @@ export const getNameFromTag = (
   return undefined;
 };
 
-export const getInstanceResponseDtoFromInstance = (
-  savedInstance: InstanceModel,
-  fetchedInstance: AWS.EC2.Instance,
-) => {
-  const instanceOption: InstanceOption = {
-    name: savedInstance.name,
-  };
-
-  const instanceInformations: InstanceInformations = {
-    id: savedInstance.instanceId,
-    type: savedInstance.type,
-    os: savedInstance.os,
-    tier: savedInstance.tier,
-    publicIp: fetchedInstance.PublicIpAddress || null,
-    privateIp: fetchedInstance.PrivateIpAddress || null,
-    securityGroup: [
-      'tcp : 80 - 0.0.0.0/0',
-      'tcp : 22 - 0.0.0.0/0',
-      'tcp : 443 - 0.0.0.0/0',
-    ],
-  };
-  const response: InstanceResponseDto = {
-    options: instanceOption,
-    informations: instanceInformations,
-  };
-
-  return response;
-};
-
 export const getInstanceResponseDtoFromInstances = (
   savedInstances: InstanceModel[],
   fetchedInstances: AWS.EC2.InstanceList[],
@@ -60,7 +27,7 @@ export const getInstanceResponseDtoFromInstances = (
     const savedInstance = savedInstances[i];
     const fetchedInstance = fetchedInstances[i][0];
 
-    const response: InstanceResponseDto = getInstanceResponseDtoFromInstance(
+    const response: InstanceResponseDto = new InstanceResponseDto(
       savedInstance,
       fetchedInstance,
     );
